Extracted permission building out of permission service
